HuggingFace/requirements.txt
tensorflow-cpu==2.12.0
tf-keras
sentence-transformers==2.2.2
gradio==4.25.0